Add movies page with data and poster assets
Introduces a new /movies route displaying a list of movies from a JSON file, including reviews and ratings. Movie poster images are moved to a dedicated assets directory, and the BurgerMenu now links to the movies page. Minor style and code cleanups are also included.

Still a work in progress, all the movies data is a placeholder